navbar works - some homepage work
